Make any binary name given to sbin'' ice be searched via preceding **/…
This allows to simply leave the ice empty, to find **/$id_as or **/$plugin,
or specify a binary name without need to give it a **/… glob, because
it is boring to do, as someone once noted (didn't find the post).

Second change is a new "1:…" option to limit the number of shims created
to only 1. So, e.g.: `sbin'1:exa*'` to have only first, shortest path
exa found be given a shim.